refactor: extract gateway lifecycle helpers
