+++ crates/daemon/src/gateway/mod.rs
@@ -5,6 +5,7 @@ pub mod api_turn;
 pub mod client;
 pub mod control;
 pub mod event_bus;
+pub(crate) mod lifecycle;
 pub mod openai_compat;
 pub(crate) mod pairing_runtime;
 pub mod read_models;
